Move shell utility functions to own module.
